requirements.txt
openai
python-dotenv
requests
arxiv
streamlit